# docs/api/order-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8086",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/cart/items/{sku}": {
            "put": {
                "tags": [
                    "shopping-cart-controller"
                ],
                "operationId": "updateItemQuantity",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCartItemQtyDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShoppingCartResponseDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "shopping-cart-controller"
                ],
                "operationId": "removeItemFromCart",
                "parameters": [
                    {
                        "name": "sku",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/cart/items": {
            "post": {
                "tags": [
                    "shopping-cart-controller"
                ],
                "operationId": "addItemToCart",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/ShoppingCartItemDto"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/order": {
            "post": {
                "tags": [
                    "order-management-controller"
                ],
                "operationId": "createOrder",
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderId"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/order/{orderId}/status": {
            "patch": {
                "tags": [
                    "order-management-controller"
                ],
                "operationId": "updateStatus",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateStatusRequestDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Accepted"
                    }
                }
            }
        },
        "/cart": {
            "get": {
                "tags": [
                    "shopping-cart-controller"
                ],
                "operationId": "getCart",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShoppingCartResponseDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "shopping-cart-controller"
                ],
                "operationId": "clearCart",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/order/me": {
            "get": {
                "tags": [
                    "order-management-controller"
                ],
                "operationId": "getCustomerOrders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderResponseDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/order/me/paged": {
            "get": {
                "tags": [
                    "order-management-controller"
                ],
                "operationId": "getCustomerOrdersPaged",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderPageResponseDto"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UpdateCartItemQtyDto": {
                "type": "object",
                "properties": {
                    "quantity": {
                        "type": "number"
                    }
                },
                "required": [
                    "quantity"
                ]
            },
            "ShoppingCartResponseDto": {
                "type": "object",
                "properties": {
                    "user_id": {
                        "type": "string"
                    },
                    "shopping_cart_items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ShoppingCartItemDto"
                        }
                    }
                }
            },
            "ShoppingCartItemDto": {
                "type": "object",
                "properties": {
                    "sku": {
                        "type": "string",
                        "minLength": 1
                    },
                    "productTitle": {
                        "type": "string",
                        "minLength": 1
                    },
                    "quantity": {
                        "type": "number",
                        "minimum": 0.0,
                        "exclusiveMinimum": true
                    },
                    "price": {
                        "type": "number",
                        "minimum": 0.0
                    },
                    "currency": {
                        "type": "object",
                        "properties": {
                            "currencyCode": {
                                "type": "string"
                            },
                            "displayName": {
                                "type": "string"
                            },
                            "symbol": {
                                "type": "string"
                            },
                            "defaultFractionDigits": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "numericCode": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "numericCodeAsString": {
                                "type": "string"
                            }
                        }
                    }
                },
                "required": [
                    "currency",
                    "price",
                    "productTitle",
                    "quantity",
                    "sku"
                ]
            },
            "OrderId": {
                "type": "object",
                "properties": {
                    "value": {
                        "type": "string"
                    }
                }
            },
            "UpdateStatusRequestDto": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    }
                }
            },
            "Money": {
                "type": "object",
                "properties": {
                    "currency": {
                        "type": "object",
                        "properties": {
                            "currencyCode": {
                                "type": "string"
                            },
                            "displayName": {
                                "type": "string"
                            },
                            "symbol": {
                                "type": "string"
                            },
                            "defaultFractionDigits": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "numericCode": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "numericCodeAsString": {
                                "type": "string"
                            }
                        }
                    },
                    "amount": {
                        "type": "number"
                    }
                },
                "required": [
                    "currency"
                ]
            },
            "OrderItemResponse": {
                "type": "object",
                "properties": {
                    "sku": {
                        "type": "string"
                    },
                    "productTitle": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "number"
                    },
                    "unitPrice": {
                        "$ref": "#/components/schemas/Money"
                    },
                    "totalPrice": {
                        "$ref": "#/components/schemas/Money"
                    }
                }
            },
            "OrderResponseDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "string"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItemResponse"
                        }
                    },
                    "status": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "subtotal": {
                        "$ref": "#/components/schemas/Money"
                    },
                    "tax": {
                        "$ref": "#/components/schemas/Money"
                    },
                    "total": {
                        "$ref": "#/components/schemas/Money"
                    },
                    "vatPercentage": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "OrderPageResponseDto": {
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderResponseDto"
                        }
                    },
                    "offset": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "limit": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "total": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "hasNext": {
                        "type": "boolean"
                    },
                    "hasPrevious": {
                        "type": "boolean"
                    }
                }
            }
        }
    }
}